automatically plot results and print worst scores after bigram score fit
